Update README.md
Rephrase to be more about us practising together as a community instead of a team of experts telling people the right way to work.

I think this is important as we will make mistakes where we suggest something that isn't great in all situations. We also want people to buy into experimenting new ways of working even if they're a bit resistant, or don't believe that what we're suggesting is the right thing to do.